requirements.txt
Django==1.5.1
amqp==1.0.12
anyjson==0.3.3
billiard==2.7.3.31
boto==2.9.9
celery==3.0.21
dj-database-url==0.2.2
django-celery==3.0.21
django-social-auth==0.7.25
django-storages==1.1.8
gunicorn==17.5
httplib2==0.8
kombu==2.5.12
oauth2==1.5.211
psycopg2==2.5.1
python-dateutil==2.1
python-openid==2.2.5
pytz==2013b
requests==1.2.3
six==1.3.0
wsgiref==0.1.2
